fix(frontend): treat "mpp" as the sole pay-as-you-go policy type
Drop the legacy "mpp_accounting" / "accounting" / "transaction" aliases
from the MPP balance card, cost extractor, and billing-policy set now that
the backend collapses them to "mpp". Delete the obsolete
TransactionPolicyContent renderer and move its shared formatCost /
formatConfigKey helpers into a new policy-format.ts consumed by the generic
policy card and the policy-item label fallback.
